Change id_image type from int to Long

+++ src/main/java/com/uade/tpo/marketplace/entity/Image.java
@@ -15,7 +15,7 @@ import java.sql.Blob;
 public class Image {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_image;
+    private Long id_image;
 
     @Lob
     private Blob image;
